+++ src/useTrackedState.js
@@ -7,7 +7,7 @@ import {
 
 import { defaultContext } from './Provider';
 
-import { useIsomorphicLayoutEffect } from './utils';
+import { useIsomorphicLayoutEffect, useAffectedDebugValue } from './utils';
 
 import { createDeepProxy, isDeepChanged } from './deepProxy';
 
@@ -50,6 +50,10 @@ export const useTrackedState = (opts = {}) => {
     const unsubscribe = subscribe(callback);
     return unsubscribe;
   }, [subscribe]);
+  if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    useAffectedDebugValue(state, affected);
+  }
   const proxyCache = useRef(new WeakMap()); // per-hook proxyCache
   return createDeepProxy(state, affected, proxyCache.current);
 };